chore: rename useGnosisSafeSdkInfo to useSafeAppsSdkInfo
Also made it use useSafeAppsSdk hook
